0-create_array.c

+++ 0x0B-malloc_free/0-create_array.c
@@ -20,7 +20,7 @@ char *create_array(unsigned int size, char c)
 
 	if (p == NULL)
 		return (NULL);
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 		p[i] = c;
 	p[i] = '\n';
